LayoutTests/imported/w3c/web-platform-tests/css-timing-1/step-timing-functions-output-expected.txt


FAIL step-start easing with input progress greater than 1 assert_equals: expected "200px" but got "100px"
FAIL step-end easing with input progress greater than 1 assert_equals: expected "100px" but got "0px"
FAIL step-end easing with input progress greater than 2 assert_equals: expected "200px" but got "0px"
FAIL step-start easing with input progress less than 0 assert_equals: expected "0px" but got "100px"
FAIL step-start easing with input progress less than -1 assert_equals: expected "0px" but got "100px"
FAIL step-end easing with input progress less than 0 assert_equals: expected "-100px" but got "0px"
